+++ CHANGELOG.md
@@ -1,9 +1,26 @@
 **Table of Contents**
 
+- [3.0.0 (development branch)](#300-development-branch)
 - [2.1.0 (2016-08-24)](#210-2016-08-24)
 - [2.0.0 (2016-08-24)](#200-2016-08-24)
 - [Previous version](#previous-version)
 
+## 3.0.0 (development branch)
+
+  - Added variable to be able to control push to Domoticz (DOMOTICZ_PUSH)
+  - Added 3 phase configuration support (using single phase by default)
+  - Added Reactive Power and Apparent Power values
+  - Added possibility to select which values should be pushed to Domoticz
+  - Added required tools check
+  - Many code improvements
+  
+  - Bash lint fixes:
+    -  Use $(..) instead of legacy `..`. [SC2006]
+    -  Double quote to prevent globbing and word splitting. [SC2086]
+
+  - TODO:
+    - TEST it with three phase system
+    - (After testing it) Add to three-phase system reactive/cosfi/apparent values
 
 ## 2.1.0 (2016-08-24)